pyproject.toml
[tool.poetry]
name = "wallet-manager"
version = "0.1.0"
description = ""
authors = ["u.veka"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
drf-yasg = "^1.21.6"
utils = "^1.0.1"
djangorestframework-simplejwt = "^5.2.2"
factory-boy = "^3.3.0"
freezegun = "^1.2.2"
coverage = "^7.2.7"
djoser = "^2.2.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"